+++ pkcs5/Cargo.toml
@@ -17,7 +17,7 @@ readme = "README.md"
 der = { version = "0.3", features = ["oid"], path = "../der" }
 spki = { version = "0.3", path = "../spki" }
 
-aes = { version = "0.7", optional = true }
+aes = { version = "0.7.3", optional = true }
 block-modes = { version = "0.8", optional = true, default-features = false }
 hmac = { version = "0.11", optional = true, default-features = false }
 pbkdf2 = { version = "0.8", optional = true, default-features = false }
