+++ pkcs5/tests/pbes2_encryption.rs
@@ -8,7 +8,7 @@ use hex_literal::hex;
 /// PBES2 + PBKDF2-SHA256 + AES-256-CBC `AlgorithmIdentifier` example.
 ///
 /// Generated by OpenSSL and extracted from the `pkcs8` crate's
-/// `tests/examples/ed25519-encpriv-aes256-sha256.der` test vector.
+/// `tests/examples/ed25519-encpriv-aes256-pbkdf2-sha256.der` test vector.
 const PBES2_PBKDF2_SHA256_AES256CBC_ALG_ID: &[u8] = &hex!(
     "305706092a864886f70d01050d304a302906092a864886f70d01050c301c0408
      79d982e70df91a8802020800300c06082a864886f70d02090500301d06096086
@@ -26,7 +26,7 @@ const ED25519_PKCS8_KEY_PLAINTEXT: &[u8] = &hex!(
 
 /// Ciphertext of Ed25519 PKCS#8 private key.
 // Extracted with:
-// $ openssl asn1parse -inform der -in tests/examples/ed25519-encpriv-aes256-sha256.der
+// $ openssl asn1parse -inform der -in tests/examples/ed25519-encpriv-aes256-pbkdf2-sha256.der
 const ED25519_PKCS8_KEY_CIPHERTEXT: &[u8] = &hex!(
     "D0CD6C770F4BB87176422305C17401809E226674CE74185D221BFDAA95069890
      C8882FCE02B05D41BCBF54B035595BCD4154B32593708469B86AACF8815A7B2B"
